backend/requirements.txt
asgiref==3.7.2
cffi==1.15.1
mysqlclient==2.2.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-rest-authtoken==2.1.4
docutils==0.20.1
pycparser==2.21
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
PyJWT==2.8.0
coverage